Write change log entries on successor writes
Change logs are initially written when a successor trie is written to
the database.

The log of the current trie (i.e. the successor's parent) is updated to
include the deletions made by the successor.  A new log entry is created
which includes the additions for the successor trie.

There can be multiple successors of a given parent trie.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>