Add StatePruningManager
The StatePruningManager is responsible fore managing a queue of state
roots and their respective block heights.

It keeps track of unique hashes and supplies an execute method which will
delete any hash below a certain depth. If there are state roots that
compute the same hash value, they will be removed so as not to cause
lost state roots for newer blocks.

Any hash unsuccessfully pruned will be put back on the queue until such
time as it can be removed.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>