Prevent deletions of leaves via duplicate log
This commit introduces the idea of a duplicate log. The duplicate log is
a record of the count of additional collisions at a given hash key.

When node is written to the database, if an entry already exists for it,
a count is updated in the duplicate log.   When a node is deleted, the
count is decremented.  Once there are no duplicates, the count is
removed entirely.  Nodes may be fully deleted at this point.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>