Add Pike project to Sabre project
The Pike Transaction Family is designed to track the identities of the
actors involved in the supply chain. These actors are agents and the
organizations they represent. The roles that the agents play within said
organizations are also tracked. This information can be used to
determine who is allowed to interact with a platform, and to what
extent they are allowed to interact with the platform.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>
Signed-off-by: Andrea Gunderson <agunde@bitwise.io>
Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>
Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>

Co-authored-by: Ryan Banks <rbanks@bitwise.io>
Co-authored-by: Andrea Gunderson <agunde@bitwise.io>
Co-authored-by: Shawn T. Amundson <amundson@bitwise.io>
Co-authored-by: Ryan Beck-Buysse <rbuysse@bitwise.io>