Add State-related traits
Add the Read, Write, and Prune traits.  These traits provide read
and commit operations, respectively, on an underlying storage mechanism.

Reading simply amounts to fetching multiple keys from state, and
returning the results in a map (a standard multi-get-style operation).

The Write and Prune operations in detail:
- commit
- compute_next_state_id
- prune

`commit`, as it implies, will actually write the values to the underlying
storage mechanism and return a new state id.

`compute_next_state_id` is effectively a dry-run of commit, providing
the same state id, but not actually writing the values to the underlying
storage mechanism.

`prune` will remove old keys that are no longer relevant prior to a
given state id.  This operation only applies to storage mechanisms that
can track change in state ids over time.

Additionally,

- Provides custom errors for the three types of state operations

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>