Add TransactionBuilder
A TransactionBuilder is used to build a TransctionPair consisting of
matching TransactionHeader, Transaction using build_pair(signer). It
can also be used to produce a Transaction with build(signer).

The builder ensures required fields are filled in and fills in fields
that can be calculated from other fields (payload hash) or by using
the signer.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>