Add feature for conditional compilation
Adds the necessary annotations to the top-level lib file to enable
the features option to the transact crate attributes.

This includes adding a 'nightly' feature to the Manifest to allow for the
project to compile using stable rust and for the project to compile with
the test crate, using the --features flag, for use with the benchmark
tests.

Signed-off-by: Shannyn Telander <telander@bitwise.io>