Replace registration callbacks with trait
Replace the registration callbacks with a single trait to collect the
registry operations of add and remove transaction family.

On the ExecutionAdapter, the two methods for adding the callbacks are
replaced with a single start method, which will trigger the adapter
to start any necessary resources for its processing of transactions.

This simplifies the use, and makes this operation atomic.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>