Remove generics from StateChange in Receipts
Removes the K, V generic types from the StateChange object from
the Receipts module, changing K and V to String and Vec<u8>,
respectively. This directly affects the implementation of the
Context Manager to handle the non-generic types, to deal with
String and Vec<u8> directly. Also changes the testing within the
Receipts and Context module to handle the type switch.

This also affects the TransactionReceipt with the removal of K & V,
and alters the use of this object within the Receipts, Scheduler,
and Context modules to reflect this change.

Signed-off-by: Shannyn Telander <telander@bitwise.io>