Add ExecutionOperationError to ExecutionAdapter
Add a Result<(), ExecutionOperationError> to all of the methods of the
ExecutionAdapter.  This allows any errors that may occur (likely due to
communication over a channel or a socket) to be propagated to the
caller.

This commit updates the two current ExecutionAdapter implementations to
propagate any errors, as well as the Executor's internal threads and
tests.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>