Fix serial scheduler join behavior
The SerialScheduler previously joined it's internal thread upon drop.
This is considered bad practice and apparently can deadlock.
Additionally, if a panic occurs during another panic which caused drop,
it causes a process abort.

Therefore, this patch moves the behavior into an explicitly called
shutdown() method. This method is part of the SerialScheduler API and
not the Scheduler API because not all schedulers may need a method to
cleanup threads. We may instead prefer to move to a model where the
scheduler, doesn't have threads in its implementation.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>