Support multi txns per batch in serial scheduler
Adds a Vec to the serial scheduler to store results of the previously
executed transactions in the current batch.

If an invalid transaction is encountered, all transactions in the
current batch (both previously executed and not yet executed) are
invalidated and added to the results.

When the transaction queue is empty, all transaction results will be
sent with the batch result.

Signed-off-by: Logan Seeley <seeley@bitwise.io>