Implement the MultiScheduler
Creates a MultiScheduler that runs multiple sub-schedulers in parallel
and verifies that they all return the same results.

To prevent the MultiScheduler from depending on the Executor, a new
trait is created to be owned by the MultiScheduler module:
SubSchedulerHandler. This trait provides a single method that the
MultiScheduler uses to pass off its sub-schedulers for execution. The
Executor implements this trait by simply passing the scheduler to its
own `execute` method.

Signed-off-by: Logan Seeley <seeley@bitwise.io>