Add logging to Sabre SDK and wasm externals
The sabre sdk now has logging macros, where if the log level
is currently supported will format the log string, write the
string to memory and call the extern function to log the message
in the Sabre transaction processor.

The log level set on the Sabre TP will also be enforced on the
smart contracts.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>