Add SchedulerError::UnexpectedNotification
Adds a new variant to the SchedulerError enum that is used when a
scheduler receives a task completion notification (from the executor)
that it was not expecting. The serial scheduler has been updated to
utilize this error when it receives a notification for a transaction
that is not the current transaction.

Signed-off-by: Logan Seeley <seeley@bitwise.io>