Propogate lock poison error in serial scheduler
Updates the serial scheduler to propogate an error rather than panic
when its shared lock is poisoned. This change includes a new
SchedulerError::Internal variant, which is used when a scheduler fails
and can't recover from the error. Also updates some of the Scheduler
methods to return Results rather than ().

Signed-off-by: Logan Seeley <seeley@bitwise.io>