Attempt to notify if serial scheduler thread ends
If the serial scheduler's core thread fails, it will attempt to send a
notification to the scheduler's calling code via the error callback; if
this fails, the error is simply logged. This change lets the calling
code know that that an internal error occurred in the serial scheduler
that it wouldn't otherwise know about until it attempts to add a batch
(at which point the core sender will fail); this allows the calling code
to take an appropriate action sooner (like shutting down the scheduler).

Signed-off-by: Logan Seeley <seeley@bitwise.io>