Serial scheduler returns None result when done
Updates the serial scheduler to return a None result when it is
finalized and all batches have been executed. This is triggered in two
ways:

1. When the scheduler is finalized, a notification is sent to the serial
scheduler's core thread; if there are no unscheduled batches and no
batch currently executing, the None result is sent.
2. If there are unscheduled batches or a batch is currently executing
when the scheduler is finalized, the scheduler will wait until all
batches have been executed, then send the None result.

Signed-off-by: Logan Seeley <seeley@bitwise.io>