Add sabre contract upload wasm flag
Add a --wasm command line argument to the sabre upload command. This
argument allows a user to specify a path to the compiled smart
contract. The value will fall back to the one specified in the manifest
file if not provided. This is useful when adding a smart contract from
a SCAR file.

Signed-off-by: Davey Newhall <newhall@bitwise.io>