Add a Redis-backed Database implementation
This commit adds a Redis-backed database implementation of the Database
trait.  It is available behind an experimental feature flag, "redis-db".

The set of tests added to the state integration tests cover the standard
operations, as well as verifying that the same hashes are produced when
a B-Tree-backed database is used.

This implementation does not currently support the
DatabaseCursor-related methods.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>