Validate custom hash lengths
Validates the custom hash lengths in the Addresser implementation
constructors. This adds a result to the constructors of the Double-
and TripleKeyHashAddressers, and makes the changes to account for
the result type. Also includes more thorough validation of the custom
input to the Addresser constructors, including a function to validate
lengths for the DoubleKeyHashAddresser. Also fixes up the error
messages in this extended validation to more accurately reflect the
cause of the error.

Signed-off-by: Shannyn Telander <telander@bitwise.io>