Add `into_payload_builder` to action builders
Add the `into_payload_builder` method to all of the SDK's
`*ActionBuilder`s. This method converts the action builder directly into
a `SabrePayloadBuilder` by building the action, adding it to a new sabre
payload builder, and returning the payload builder. This change
simplifies the process of building a payload from scratch.

Signed-off-by: Logan Seeley <seeley@bitwise.io>