Look for .scar in multiple paths
Updates the `SmartContractArchive::from_scar_file` constructor to take
multiple paths and find the .scar with the highest matching version in
all of the paths.

Previous to this change, if a user of this library wanted to do this,
they would have to make multiple calls to the `from_scar_file`
constructor, filter out any failures (in the case that a matching .scar
wasn't in one of the directories) and compare the resulting versions
manually (rather than letting the library handle this as was intended).

Signed-off-by: Logan Seeley <seeley@bitwise.io>