Enable SQLite WAL mode
Add a builder parameter to enable SQLite's "Write-Ahead Log" journal
mode.  This mode has significant performance improvements over the
standard mode.  While many of the disadvantages of this mode are
unlikely to effect transact and its use-cases, WAL mode cannot be used
with a database file on a networked file system.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>