Add `Scheduler::shutdown` method
Adds the `shutdown` method to the `Scheduler` trait. This method is
required to clean up background threads (like in the serial scheduler).

The existing implementations of `shutdown` for the serial and multi
schedulers are converted to trait methods in this commit.

Also fixes a comment typo in the `shutdown` implementations.

Signed-off-by: Logan Seeley <seeley@bitwise.io>