Update `Scheduler::cancel` to abort execution
Updates the `Scheduler::cancel` method's docs to indicate that the
method should abort and return any batches that are currently executing.
The purpose of this change is to ensure that when a scheduler is
cancelled, the calling code can retrieve all scheduled batches in a
timely manner. It is possible for a batch to take a long time to
execute; without aborting currently-executing batches, the calling code
has no way of getting those batches back if it needs to drop the
scheduler immediately.

This commit includes an update to the `SerialScheduler` implementation
to properly abort and return its pending batch if it has one. It also
removes the tests that verify an `UnexpectedNotification` is sent when
an unexpected task completion notification is received; this is required
because there is now a valid situation (aborted batch) where a
notification may be received for a transaction that the scheduler is not
waiting for.

Signed-off-by: Logan Seeley <seeley@bitwise.io>