Send `None` when cancelling finalized schedulers
Updates the scheduler docs to indicate that cancelling a finalized
scheduler will result in the scheduler sending a `None` batch execution
result. This signals to the calling code that no more batches will be
received, and allows the calling code to always be able to determine
that the scheduler has fully completed just by looking at the batch
results.

Updates the serial scheduler to return this `None` result, and updates
the multi scheduler to ensure that the `None` result is always returned
when no more results will be produced.

Signed-off-by: Logan Seeley <seeley@bitwise.io>