Take non-mut self in `create_scheduler`
Updates the `SchedulerFactory::create_scheduler` method to take self as
non-mutable, which removes the need for the consumer to borrow it as
mutable. A factory should be considered stateless from the consumer's
perspective; this means that the factory should not change in any way
that is visible to the consumer, and the interface itself should not
have mutability built-in.

The multischeduler's factory requires interior mutability because of the
mutability in the `SubSchedulerHandler` trait. This trait was designed
as an abstraction of the `Executor`; the assumptions about mutability in
the `SubSchedulerHandler` trait should be reconsidered in the future.

Signed-off-by: Logan Seeley <seeley@bitwise.io>