Replace signing module with cylinder
Replaces the `transact::signing` module with the `cylinder` library,
which provides a similar interface and implementations.

The use of the `HashSigner` has been replaced by the secp256k1
implementation that is provided by cylinder; wherever the hash signer
was used, transact now creates a secp256k1 signer with a random key.
This lead to two public API changes:

* The `families::command::make_command_transaction` function now takes a
  signer as an argument; this allows the calling code to control the key
  that's used, which may be desirable in some cases.
* The `XoTransactionWorkload` and `XoBatchWorkload` now have a single
  constructor, `new`, that takes both a workload seed and a signer as
  optional arguments; this is a simplification of the interface that
  allows the calling code to pick the seed and key where needed, while
  providing the same ease-of-use as before with the `default` method.

Signed-off-by: Logan Seeley <seeley@bitwise.io>