Add ExecutionTaskSubmitter
This change adds the ExecutionTaskSubmitter struct, which is returned by
the Executor.  This struct can be sent created or cloned and sent to
threads. It replaces the internals of Executor::execute with
ExecutionTaskSubmitter::submit, which sends the task iterators and
completion notifiers to the Executor's command thread.

The existing Executor::execute function remains, but just forwards to
an ExecutionTaskSubmitter instance.

The sawtooth compatibility tests and the simple_xo example have been
updated to make use of this new API.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>