Fix a bug where None is never returned on SerialScheduler completion
If the scheduler was finalized while there was still unscheduled
batches, None would not be returned. When try_schedule_next
was called, the scheduler would not be ready. Since the task
iterator was empty, next will never be true and try_schedule_next
would never be called again.

This commit checks to see if the all batches have been scheduled
and returns None if finalized before checking if the scheduler
is ready. This will allow the None to always be returned when
complete.

This commit also fixes a bug where if there are no unscheduled
batches the SerialScheduler would always shutdown, even
if not finalized.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>