Fix serial scheduler dropped receiver bug
Fixes a bug where the internal receiver of the serial scheduler is
dropped because the internal thread has shutdown. This error occurs when
the scheduler has been finalized and has no more batches, and then after
some time the scheduler is cancelled; the receiving end has been
dropped, so the cancellation message cannot be sent.

Signed-off-by: Logan Seeley <seeley@bitwise.io>