Update shutdown process for serial scheduler
This commit adds the `try_shutdown` method to the serial scheduler's
core thread struct. This method is now called in all cases where the
scheduler's state has changed in such a way that it may be ready to
shutdown.

Previously, the shutdown process was handled a bit differently in
various places. This change consolidates the logic into a single method
that is called consistently everywhere. By using the same logic for
shutdown everywhere that the scheduler could shutdown, we ensure that
the scheduler shuts down properly every time, which includes notifying
all interested components of the shutdown.

One shutdown task that was not done reliably was sending a `None` to the
execution task iterator if a task is expected. This is now done every
time the scheduler shuts down to ensure that the iterator does not block
indefinitely when it is waiting for a task. This was previously only
done in the `CoreMessage::Next` handler, but is now handled by
`try_shutdown`.

Also removes the `bool` returned by `try_schedule_next`, since it no
longer contains the logic for shutting down.

Signed-off-by: Logan Seeley <seeley@bitwise.io>