Fix race condition in serial scheduler test
Fixes a race condition in the `test_serial_scheduler_ordering` test that
occasionally cause the test to fail. When the `tx.send` was after the
task notifier `notify` call, sometimes the message wasn't sent from the
spawned thread before the main test thread got to the `rx.try_recv`.
This caused the test to fail.

Signed-off-by: Logan Seeley <seeley@bitwise.io>