Make cylinder dependency optional
When compiling with --no-default-features, we want the code to compile
to the wasm-unknown-unknown target. Cylinder doesn't currently support
the wasm-unknown-unknown target, so it is necessary to make it optional
by default.

This defines a set of features to accomplish this. The first are
protocol-batch and protocol-transaction, which guard the code which
directly depends on Cylinder. The next are a layer up: execution,
handler, and scheduler; these all require protocol-batch and
protocol-transaction in their APIs/implementations.

These features are added to "default" currently for compatibility but
long-term we may want to move them to stable.

In addition, the "log" feature is marked as optional and set as
a dependency for the features that need it. This prevents compiler
warnings when those features are not enabled.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>