Fix guards around family-command
Previously we enabled family-command for tests; however, this was an
incorrect approach as it enables the code in situations where tests are
being run but family-command feature isn't intentionally enabled.

This removes that approach and replaces it with a check on the execution
adapters tests which require family-command be enabled. This may cause
the tests to run less often but it is more correct.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>