Add WorkloadRunner to run BatchWorkloads
The WorkloadRunner makes it easy to run multiple workloads
against different groups of URLS. The workload groups can
be started and removed dynamically, allowing for more
complicated workload test.

This required updating the TransactionWorkload and
BatchWorkload traits to require they can be Send because
the workloads will be run in their own thread.

This is behind an experimental feature "workload-runner"

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>