Add initial transact workload command to CLI
This command allows a workload to be submitted against
groups of targets, a collection of URL that will share state
(aka need to work of the same workload).

Currently smallbank is the only supported workload type and
requires that the targets are running Sawtooth Sabre and the
smallbank sabre smart contract has already been configured.

The workload commands also assumes that the targets support
Cylinder JWTs for REST API authentication.

The workload will run until a ctrl-c, then the workloads will
be shutdown.

This is behind an experimental features: workload

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>