Make smallbank workload re-runnable
Before the number of accounts were generated with
client ids in order from 0 to number account requested.
This made it impossible to rerun the workload against
the same network with out having at least the number
of accounts invalid transactions. Instead, the client
id are now randomly generated.

If a seed is provided, the output will be generated the
same way, causing invalid transactions. To be rerun
without the invalid transactions, run again with a
different or no seed to create new accounts.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>