Add missing check for if source savings balance is zero
The amalgamate transaction should check to see if the savings
balance of the source account is 0 and return an invalid
transaction if it is.

This was specified in the transaction family specification
but was missed in the implementation.

Without this check the transaction would act as an unexpected
noop transaction.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>