Add namespace to command family
Add `namespaces` to the `CommandTransactionHandler` struct to allow for
namespace to be set when running the smart contract with scabbard.

Add a `get_command_prefix` method which will return the first 6
characters of the sha512 hash of "command" and use it to set the
`namespaces` field when a new CommandTransactionHandler is created.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>