Add `slow_rate` method
This function uses the start time and total number of submitted batches
to calculate the effective submission rate and the amount of time to
sleep based on that rate. After sleeping for the calculated amount of
time it will attempt to re-submit the given batch and loop until
successful, sleeping between each submission.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>