Move Node struct to parent mod
This change moves the Node struct to the parent merkle module, such that
it can be shared across implementations.  This will ensure that the
identical hash values are produced for the key-value implementations as
well as any future implementations.

Additionally, this changes the error type for Node serialization and
deserialization to InternalError, in order to make the API more
general-purpose.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>