CommandTransactionWorkload & CommandBatchWorkload
Create a `CommandTransactionWorkload` struct that contains a signer
and `CommandGeneratingIter`. Implement
`TransactionWorkload` for `CommandTransactionWorkload`, the
`next_transaction` method creates a `TransactionPair` out of the command
proto created by the generator.

Create a `CommandBatchWorkload` struct that contains a signer and a
`CommandTransactionWorkload`. Implement `BatchWorkload` for
`CommandBatchWorkload`, the `next_batch` method creates a `BatchPair`
out of a `TransactionPair` created by the `CommandTransactionWorkload`.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>