Add CommandGeneratingIter
Create a `CommandGeneratingIter` struct. Implement Iterator for
`CommandGeneratingIter` so that it can be used to continuously create
command family commands. The `CommandGeneratingIter` is initialized
with a vector of 100 addresses that adhere to the command family
namespace, these addresses are randomly selected to be used for the
commands. The `CommandGeneratingIter` keeps track of addresses
that have been set, this is because the `delete_state` command will
return an error if it attempts to delete a state that has not been set.
A `delete_state` command will only be generated if there are existing
set addresses.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>