Update `slow_rate` to handle control-c
Previously if control-c was pressed while the `slow_rate` method was
attempting to resubmit a batch the workload would not stop until
`slow_rate` had returned. This commit adds a `receiver` argument to
the `slow_rate` method, the method uses the receiver to check if a
shutdown message has been sent to the worker each time it loops. The
`slow_rate` method is updated to return a boolean representing
whether or not a shutdown message was sent. If true is returned to the
loop from which `slow_rate` is called, the loop will break and the
workload will stop.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>