Add MerkleUpdateIndexOperation
This change adds an operation method to update the index based on a
state root hash, a parent state root hash and the set of changed leaves.
The index is created using slowly-changing dimensions (type 2).  The
index only tracks a linear set changes.  Any forks are removed from the
index and the leaves must be queried via the tree node table.

It includes two allow(dead_code) metadata tags, which will be removed in
a future commit that will make use of the code.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>