Add SqlMerkleState
This struct provides the SQL implementation of MerkleState, via the
state::Read, Write traits, as well as the MerkleRadixLeafReader trait.

Likewise, the integration tests have been added for this implementation.
This includes a new test that allows for comparisons between MerkleState
implementations only using the Read and Write traits.

It omits the state::Prune trait for now (this will be implemented in a
future commit).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>