Add merkle_radix_tree table
This change adds a merkle_radix_tree that allows for named trees.  Each
record in the existing merkle_radix tables includes a link to a given
tree. In the case of the merkle_radix_tree_node table, the primary key
for a node becomes a composite key of (hash, tree_id).

All code has been updated to include a reference to tree_id 1, the
default tree.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>