Create specific models and tables for sqlite
This change alters the schema and models for the merkle_radix_tree_node
table to create an explicit sqlite table and structs.

The underlying table name is the same, but the name in Rust/diesel
explicitly includes `sqlite_` in the name.  Likewise the models for the
table have been moved entirely into the `models::sqlite` submodule.

The tests have been updated use these new tables and the sqlite nature
of each test has been called out explicitly.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>