Add run_postgres_test test runner
This change adds a test runner method for postgres tests.  It uses a DB
url configured via an environment variable. The migrations are applied
before the test is run.  Once the test is complete (regardless of
success) all the migrations are rolled back, returning the DB to the
original state.

Tests using this runner are run serially, enforced by a lock.

This function is only available under the test profile, as well as under
the "state-merkle-sql-postgres-tests" feature.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>