Replace checked_sub with saturating_sub
When I originally wrote this I wanted to use saturating_sub because it
mapped nicely to the requirements but it was one version of Rust off.

Its now stable.

The code removed was copied from the saturating_sub pull request to the
std library so for now there is no real change at all. Going forward relying
on the std lib implementations is preferable to just using unpatched
code snippets copied from pull requests.

Signed-off-by: Caleb Hill <hill@bitwise.io>