Replace get_leaves by index with direct query
This change replaces the get_leaves index query with a recursive query
in order to look up the leaves.  This ensures correctness of the tree,
without needing to determine if the index currently exists or not.

This commit also updates the tests to match the new functinality, though
in a simplified manner (complex uses are covered by the integration
tests).

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>