Remove StateDatabaseError::InternalError variant
This change removes the StateDatabaseError InternalError variant, as
this is a backwards-incompatible change.

In order to remove this, the original Node implementation has been
restored in the kv module, the state::merkle::node::Node implementation
has been reduced to only what the SQL implementation requires, and is
guarded by the "state-merkle-sql" feature.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>