Allow migrations to be run against a conn
This change allows the user to run the migrations directly against a
reference to a diesel connection.  It accomplishes this by re-exporting
the database-specific migration functions directly from the migration
module.

This allows migrations to be run in instances where the caller does not
have access to a pool or the connection string, and therefore cannot use
a `Backend` instance via the `MigrationManager`.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>