Add `check_batch_status`
Add a method `check_batch_status` that takes the HashMap containing
the batch status links and expected batch results and loops through the
links checking if the batch status matches what was expected. If the
status link returns the expected result the link is removed from the
list so that it will not be checked again. If the status is pending
the link will be left in the list so that it can be checked again on
the next iteration.

`check_batch_status` is called at the end of the loop if
`get_batch_status` is true.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>