Add `shutdown_signaler` and `wait_for_shutdown`
Add a struct `WorkerShutdownSignaler` which has a vector of tuples
containing a worker id and its sender. `WorkerShutdownSignaler` has a
function `signal_shutdown` which loops through the list of senders
sending a `ShutdownMessage` to each.

Add a `shutdown_signaler` function to `WorkloadRunner` which returns a
`WorkerShutdownSignaler`.

Add a `wait_for_shutdown` function to `WorkloadRunner` which joins all
worker threads and blocks until the threads have finished.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>