Change ctrlc handling
Change ctrlc handling to use the `shutdown_signaler` to shutdown worker
threads rather than using AtomicBool to track the state of the
WorkloadRunner.

Call `wait_for_shutdown` after setting the ctrlc handler instead of
using the loop + AtomicBool.

This change allows for the workload to shutdown if any errors occur in
the worker thread. Previously, if an error occurred in the loop in the
worker's thread, the workload would stop submitting batches but the
program would stall because the outer loop was not broken until a SIGINT
was recieved.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>