Replace exit with break
Remove the exit that is called when a batch's status does not match
the expected status and replace it with a break. Using break to
leave the loop and return the worker is safer than using exit
because when exit is called no destructors on the current stack or
any other thread’s stack will be run.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>