Update `slow_rate` shutdown at endtime
Update the `slow_rate` method to set the shutdown variable to `true`
when returning because the end time has been reached. Without this
update, if the end time was reached while `slow_rate` was attempting to
resubmit a batch, an inaccurate error was being logged "Failed to get
batch status link", because shutdown was set to false.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>