Add helper `signal_batch_status_checker_shutdown`
Add a helper function to send a `ShutdownMessage` to a worker's
batch status checker if `check_batch_status` is true. A worker will not
have a batch status checker if `check_batch_status` is not true.

Call `signal_batch_status_checker_shutdown` everywhere that the worker
thread breaks to complete the thread.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>