Create a `RequestLogger` to print batch updates
Add a `request_logger` module that contains `RequestLogger`.
`RequestLogger` has a thread that loops through a list of
`HttpRequestCounter`s and prints the information stored in them,
sleeping for a specified amount of time between reports.

Add accessor methods to `HttpRequestCounter` so that values in the
struct can be retrieved/modified in the `RequestLogger`

Signed-off-by: Isabel Tomb <tomb@bitwise.io>