Update `CommandGeneratingIter` to remove panics
This commit adds a `CommandType` enum to represent the different
commands that the `CommandGeneratingIter` can create and implements
`Distribution` using the `CommandType` enum as the type parameter. This
change makes it possible to call rng.gen() and have it generate a
random `CommandType`.

These changes allow for all panics to be removed.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>