Promote ContextId to struct
This change promotes the ContextId type to a struct from a simple type
definition.  This resolves an issue where a context id cannot be
safely logged if its bytes are not valid UTF-8, which can only be safely
done by knowing what the bytes are.  As the bytes were created internally
by Context, calling structs, such as the ContextManager have no
visibility into what the type is.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>