Drop contexts in serial sched when txns complete
This change drops the appropriate contexts when a transaction is
complete or the schedule is canceled.

When a transaction is complete and there is a next transaction, the
previous context is dropped after the new context is created.

It updates the current transaction to include its associated context ID
so that the context can be dropped on cancellation.  Otherwise, the
context would get leaked.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>