Add `BatchReader` and `TransactionReader` traits
Add a `transaction_reader` and a `batch_reader` module which
contain the  traits `BatchReader` and `TransactionReader`. These
traits provide an API for reading batches and reading transactions
from a source.

Add implementations of these traits in transaction_reader::protobuf
and batch_reader::protobuf which read length-delimited protobuf
from a file.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>