Allow dead code
Add `#[allow(dead_code)]` to the `ValidTransaction` and
`InvalidTransaction` structs and make fields not pub. These
structs are printed out in an error message if an unexpected
batch result is encountered, removing the fields that are
being reported by clippy as unused would remove important
information from the error messages.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>