Remove Journal settings from sqlite tests
This change removes the inclusion of WAL mode in the tests.  This change
was made to fix the multithread_read test, though the use of WAL doesn't
guarantee successful multi-threaded use.

As the sqlite implementation of the MerkleRadixStore now uses the
correct single writer method for executing changes, this test properly
operates without any underlying PRAGMA settings.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>