Alter column type for leaf address
Prior to this change, the column type for merkle_radix_leaf.address was
specified as "STRING". As this type is not one of the types
automatically treated as TEXT, it defaults to being treated as a NUMERIC
value.  With certain addresses, where all characters were valid decimal
digits with a leading `0`, the leading zero would be dropped. This would
create instances where the address value would not be returned, but not
with the same address queried.

Changing the column type to "TEXT" fixes the problem.

The migration added will also correct all address that are missing a
leading zero.

This change also adds several tests for both SQLite and Postgres to
guard against this issue in the future, while also exercising the
list_leaves trait function implementations of each.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>