Fix transmute_num_to_bytes lint
This change fixes a lint introduced with the Rust 1.58 release. The lint
warns of using transmute to convert a number to an array of bytes.  As
this is an unsafe operation, it is a very good lint to follow.  It can
be replaced with a std library function introduced in Rust 1.32.

See
https://rust-lang.github.io/rust-clippy/master/index.html#transmute_num_to_bytes
for more details.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>