Add state root hash to deletions
This change adds the state root hash to the set of deletions from the
prior tree.  Minimally, the state root should change from the previous
root (i.e. a new address has been added to the tree that creates a new
branch on the root).

This will result in at least one record in the
merkle_radix_change_log_addition table, which will minimally indicate
that there is at least one successor.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>