Add Execute and WriteExclusiveExecute traits
This change adds two traits for the style of execution a backend may
support.

The Execute trait supports execution where no indication of whether the
connection is going to be used for reads or writes is required.

The WriteExclusiveExecute trait allows the user to indicate if the
connection is used for reading or writing, where the later may have
further restrictions on how the connection is used.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>