Split Read + Sync + Send into two traits
This change splits the Read trait into two traits: the base Read trait
that specifies the types (state ID, key, value) and operations, and a
SyncRead trait which requires the implementer to be Sync + Send, with
the clone_box function.

This split will allow for implementations of Read to be temporary, based
on references to databases or their connections.  Without this change,
it is not possible to create such a state struct.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>