Remove Clone constraint for Backend
This change removes the Clone constraint on the backend generic
parameter for SqlMerkleState.  It implements Clone on the struct only
when the provided parameter implements clone.

This allows for backends that may not be clonable.  For example, a
backend with reference to a single connection should not necessarily be
cloned.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>