Add PreparedStatement
This change adds a PreparedStatement, providing the same API as
sql_query but allowing the statement to be cached.  As the statement is
not constructed by external operations, it is safe to cache.

It provides a significant performance boost with postgres.

This code is derived from the diesel-rs implementation of SqlQuery.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>